+++ controllers/contacts/index.js
@@ -1,0 +1,14 @@
+const getAll = require('./getAll')
+const getById = require('./getById')
+const add = require('./add')
+const putById = require('./putById')
+const delById = require('./delById')
+
+
+module.exports = {
+    getAll,
+    getById,
+    add,
+    putById,
+    delById
+}
