+++ models/contacts.json
@@ -58,5 +58,11 @@
     "name": "Roger2101",
     "email": "drg@mail.com",
     "phone": "(233) 738-2360"
+  },
+  {
+    "id": "bf33e323-22a8-4fe7-bd62-f5f587dfc4e5",
+    "name": "Alex12121",
+    "email": "alex@mail.com",
+    "phone": "08524344887"
   }
 ]
