+++ schemas/contacts.js
@@ -1,0 +1,9 @@
+const Joi = require('joi')
+
+const schema = Joi.object({
+  name: Joi.string().min(3).max(30).required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string().min(10).max(15).required(),
+})
+
+module.exports = schema
